backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5.0
django-filter>=24.3,<25.0
django-celery-beat>=2.7,<3.0
django-celery-results>=2.5,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<24.0
celery>=5.4,<6.0
redis>=5.2,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
openai>=1.50,<2.0
tiktoken>=0.7,<1.0
PyPDF2>=3.0,<4.0
python-docx>=1.1,<2.0
python-magic>=0.4,<1.0
Pillow>=10.4,<11.0
dj-database-url>=2.2,<3.0
whitenoise>=6.7,<7.0
pydantic>=2.9,<3.0
httpx>=0.27,<1.0
uvicorn>=0.30,<1.0
python-json-logger>=2.0,<3.0
django-storages>=1.14,<2.0
